src/requirements.txt
# Dependências principais
streamlit
pandas
requests
python-dotenv

# Tradução automática
deep-translator
